+++ widgets/date.lua
@@ -1,0 +1,16 @@
+local wibox = require("wibox")
+local awful = require("awful")
+local cal = require("utils.cal")
+
+dateicon = wibox.widget.imagebox()
+dateicon:set_image(beautiful.widget_date)
+-- Initialize widget
+datewidget = wibox.widget.textbox()
+-- Register widget
+vicious.register(datewidget, vicious.widgets.date, "%b %d, %R", 60)
+-- Register buttons
+datewidget:buttons(awful.util.table.join(
+  awful.button({ }, 1, function () exec("pylendar.py") end)
+))
+cal.register(datewidget)
+
