+++ widgets/system.lua
@@ -1,0 +1,58 @@
+local wibox = require("wibox")
+local awful = require("awful")
+
+-- {{{ CPU usage and temperature
+cpuicon = wibox.widget.imagebox()
+cpuicon:set_image(beautiful.widget_cpu)
+-- Initialize widgets
+cpugraph  = awful.widget.graph()
+tzswidget = wibox.widget.textbox()
+-- Graph properties
+cpugraph:set_width(40):set_height(14)
+cpugraph:set_background_color(beautiful.bg_widget)
+cpugraph:set_color(gradient_colour)
+-- Register widgets
+vicious.register(cpugraph,  vicious.widgets.cpu,      "$1")
+vicious.register(tzswidget, vicious.widgets.thermal, " $1C", 19, {"thermal_zone1", "sys"})
+-- }}}
+
+-- {{{ RAM usage
+ramicon = wibox.widget.imagebox()
+ramicon:set_image(beautiful.widget_mem)
+ramgraph = wibox.widget.textbox()
+vicious.register(ramgraph, vicious.widgets.mem, "$1%", 1)
+--- }}}
+
+-- {{{ Battery state
+baticon = wibox.widget.imagebox()
+baticon:set_image(beautiful.widget_bat)
+-- Initialize widget
+batwidget = wibox.widget.textbox()
+-- Register widget
+vicious.register(batwidget, vicious.widgets.bat, "$2%", 61, "BAT0")
+-- }}}
+
+-- {{{ File system usage
+fsicon = wibox.widget.imagebox()
+fsicon:set_image(beautiful.widget_fs)
+-- Initialize widgets
+fs = {
+  r = awful.widget.progressbar(), h = awful.widget.progressbar()
+  --s = awful.widget.progressbar(), b = awful.widget.progressbar()
+}
+-- Progressbar properties
+for _, w in pairs(fs) do
+  w:set_vertical(true):set_ticks(true)
+  w:set_height(14):set_width(5):set_ticks_size(2)
+  w:set_border_color(beautiful.border_widget)
+  w:set_background_color(beautiful.fg_off_widget)
+  w:set_color(gradient_colour)
+  --w.widget:buttons(awful.util.table.join(
+  --  awful.button({ }, 1, function () exec("rox", false) end)
+  --))
+end -- Enable caching
+vicious.cache(vicious.widgets.fs)
+-- Register widgets
+vicious.register(fs.r, vicious.widgets.fs, "${/ used_p}",            599)
+vicious.register(fs.h, vicious.widgets.fs, "${/boot used_p}",        599)
+-- }}}
